requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-google-genai
langchain-chroma
chromadb
sentence-transformers
google-generativeai
python-dotenv
pypdf
IPython
